Change to u64

+++ src/lib.rs
@@ -1,14 +1,14 @@
-fn is_prime(n: u32) -> bool {
+fn is_prime(n: u64) -> bool {
     if n == 1 {
         false
     } else {
-        let max = (n as f32).sqrt() as u32;
+        let max = (n as f32).sqrt() as u64;
         !(2..max + 1).any(|x| n % x == 0)
     }
 }
 
-pub fn get_prime_list(n: u32) -> Vec<u32> {
-    let mut primes: Vec<u32> = Vec::new();
+pub fn get_prime_list(n: u64) -> Vec<u64> {
+    let mut primes: Vec<u64> = Vec::new();
     if n == 0 {
         return Vec::new();
     } else if n == 1 {
@@ -23,7 +23,7 @@ pub fn get_prime_list(n: u32) -> Vec<u32> {
     primes
 }
 
-pub fn sum(n: u32) -> Option<u32> {
+pub fn sum(n: u64) -> Option<u64> {
     let prime_list = get_prime_list(n);
     match prime_list.is_empty() {
         true => None,
